backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
networkx==3.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
